% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{AdjustCN}
\alias{AdjustCN}
\title{Corrects CN and filters cells based on AdjustPerCell output}
\usage{
AdjustCN(
  scCN,
  PerCell,
  Basename_leves = NULL,
  Group_leves = NULL,
  Chr_filter = NULL
)
}
\arguments{
\item{scCN, }{scCN dataframe created by CallCNV}

\item{PerCell, }{PerCell dataframe created by AdjustPerCell}

\item{Basename_leves, }{Sample basenames order (optional)}

\item{Group_leves, }{Group basenames order (optional)}

\item{Chr_filter, }{filters chromosomes to keep in the analysis (optional)}
}
\value{
tibble
}
\description{
Corrects CN and filters cells based on AdjustPerCell output
}
