% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{AdjustPerCell}
\alias{AdjustPerCell}
\title{Applies diagnostic settings to PerCell dataframe}
\usage{
AdjustPerCell(PerCell, Settings, Basename_leves = NULL, Group_leves = NULL)
}
\arguments{
\item{PerCell, }{PerCell dataframe created by CallCNV}

\item{Settings, }{Setting dataframe created by diagnostic}

\item{Basename_leves, }{Sample basenames order (optional)}

\item{Group_leves, }{Group basenames order (optional)}
}
\value{
tibble
}
\description{
Applies diagnostic settings to PerCell dataframe
}
