% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_subpop.R
\name{Assign_subpop}
\alias{Assign_subpop}
\title{Splits single cell CN dataframe  based on identified sub-populations}
\usage{
Assign_subpop(scCN, Supbop, group = NULL, basename = NULL)
}
\arguments{
\item{scCN, }{single cell CN data frame produced by Kronos}

\item{Supbop, }{Sub-population grouping created by Kronos}
}
\value{
list
}
\description{
Splits single cell CN dataframe  based on identified sub-populations
}
