% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{AverageVariability}
\alias{AverageVariability}
\title{Calculates the proportion of replicated bins in function of time}
\usage{
AverageVariability(Variability)
}
\arguments{
\item{Variability, }{dataframe created by either TW_RTAnnotation or TW_GenomeAnnotation}
}
\value{
ggplot element
}
\description{
Calculates the proportion of replicated bins in function of time
}
