% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{BackGround}
\alias{BackGround}
\title{Calculates median G1/G2 profile}
\usage{
BackGround(G1_scCN)
}
\arguments{
\item{G1_cells, }{scCN dataframe from Rebin function}
}
\value{
tibble
}
\description{
Calculates median G1/G2 profile
}
