% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{BamMetrics}
\alias{BamMetrics}
\title{recover metrics from bam files}
\usage{
BamMetrics(Bamfile, isPE = T)
}
\arguments{
\item{Bamfile, }{a bam file directory}

\item{isPE, }{logical value indicating whether reads are PE (TRUE) or SE (FALSE)}
}
\description{
recover metrics from bam files
}
