% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{binning}
\alias{binning}
\title{Calculates genomic bins with GC content and mappability}
\usage{
binning(
  RefGenome,
  bowtie2_index,
  bin_size = 20 * 10^3,
  read_size = 40,
  fragment_size = 200,
  paired_ends = F,
  directory_to_bamfiles = NULL,
  tmp_dir = file.path(getwd(), "bins"),
  upper_mappability_th = 1.5,
  lower_mappability_th = 0.8,
  black_list = NULL,
  coverage = 1,
  errorRate = 10^-3,
  chr_prefix = "chr",
  chr_range = NULL,
  cores = 1,
  return_estimated_param = F
)
}
\arguments{
\item{RefGenome, }{path to .fa file}

\item{bowtie2_index, }{bowtie2 index}

\item{bin_size, }{size of the final bin in bp: defailt = 20 000}

\item{read_size, }{size of a read in bp: default= 40bp}

\item{fragment_size, }{size of a fragment in bp if paired_ends=T: default= 200bp}

\item{paired_ends, }{logical, whether reads have to be PE or SE: default= False}

\item{directory_to_bamfiles, }{path to bam file folder from which estimate read_size, fragment_size and paired_ends.}

\item{tmp_dir, }{path where to create a temporary folder: default= bins in the wd}

\item{upper_mappability_th, }{max mappability above which a bin is considered unsuitable for the analysis}

\item{lower_mappability_th, }{min mappability below which a bin is considered unsuitable for the analysis}

\item{black_list, }{path to a blacklist file}

\item{coverage, }{an integer corresponding to the simulated coverage}

\item{errorRate, }{simulated sequencer percentage error rate}

\item{chr_prefix, }{prefix identifying a chromosome in the .fa file: default='chr'}

\item{chr_range, }{chomosomes to considere in the analysis. Use ':' to identify a range and ',' for isolated chromosomes. default='1:22'}

\item{cores, }{number of cores to use}

\item{return_estimated_param, }{logical value. If True and directory_to_bamfiles has been passed it returns a list wit the estimated parameters and the calculated bins}
}
\value{
dataframe
}
\description{
Calculates genomic bins with GC content and mappability
}
