% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNV_functions.R
\name{CallCNV}
\alias{CallCNV}
\title{Calls CNV over genomic bins defined by binning}
\usage{
CallCNV(
  directory,
  bins,
  chrom_size,
  basename = "Exp",
  group = NULL,
  tmp_dir = file.path(getwd(), "CNV"),
  min_n_reads = 2e+05,
  mim_mean_CN_accepted = 2,
  max_mean_CN_accepted = 8,
  ploidy = NULL,
  chr_prefix = "chr",
  chr_range = NULL,
  cores = 1
)
}
\arguments{
\item{directory, }{Single cell Bamfiles directory}

\item{bins, }{bins produced by binning}

\item{chrom_size, }{a dataframe (chr=chrom name, size=chrom size)}

\item{basename, }{sample name}

\item{group, }{experimental group name}

\item{tmp_dir, }{temporary folder.}

\item{min_n_reads, }{Min n of reads to keep a cell in the analysis}

\item{mim_mean_CN_accepted, }{Min mean CN accepted as result}

\item{max_mean_CN_accepted, }{Max mean CN accepted as result}

\item{ploidy, }{user extimated ploidy}

\item{chr_prefix, }{Chromosome prefix, if there is no prefix use NULL}

\item{chr_range, }{Chromosomes to consider in the analysis (example 1:5,8,15:18,X)}

\item{cores, }{Number of cores to use}
}
\value{
list of dataframes
}
\description{
Calls CNV over genomic bins defined by binning
}
