% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{Cluster_enrichment}
\alias{Cluster_enrichment}
\title{Gene ontology on RT clusters using clusterprofiler}
\usage{
Cluster_enrichment(
  clusters,
  genes,
  orgDB,
  ontology = c("BP", "MF", "CC", "ALL"),
  showCategory = 10,
  qvalueCutoff = 0.05,
  ...
)
}
\arguments{
\item{clusters, }{clusters from RT_clustering}

\item{genes, }{either a dataframe containing chr,start,end,gene column, where gene is an ENSEMBL identifier or the path to a gtf file}

\item{orgDB, }{OrgDb.}

\item{ontology, }{One of "BP", "MF", and "CC" sub-ontologies, or "ALL" for all three.}

\item{showCategory, }{number of categories to show per cluster}

\item{qvalueCutoff, }{cut-off for qval}

\item{..., }{other parameters to pass to enrichGO}
}
\value{
ggplot
}
\description{
Gene ontology on RT clusters using clusterprofiler
}
