% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{Decompress_if_needed}
\alias{Decompress_if_needed}
\title{Checks if files are compressed and decompressed them in a destination folder returning the full path of the decompressed file or original file if no decompression was needed}
\usage{
Decompress_if_needed(File, outdir = NULL)
}
\arguments{
\item{File, }{file to test}

\item{outdir, }{output directory}
}
\description{
Checks if files are compressed and decompressed them in a destination folder returning the full path of the decompressed file or original file if no decompression was needed
}
