% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{FastqToBam}
\alias{FastqToBam}
\title{Trims and maps reads based on Rbowtie2}
\usage{
FastqToBam(
  bowtie2_index,
  File1,
  File2 = NULL,
  adapters_list = NULL,
  outputdir = file.path(getwd(), "FastqToBam"),
  cores = 1,
  trim = T,
  phred33 = T,
  Read_min_size_after_trimming = 25
)
}
\arguments{
\item{bowtie2_index, }{path to bowtie2 index}

\item{File1, }{list of paths to R2 fastq files for PE}

\item{adapters_list, }{list of adapters to trim, if not provided general illumina adaters will be used}

\item{outputdir, }{output path}

\item{cores, }{number of parallel jobs}

\item{trim, }{logical value, if False reads won't be trimmed}

\item{phred33, }{logical value, if False phred64 will be used}
}
\description{
Trims and maps reads based on Rbowtie2
}
