% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{find_known_sequences}
\alias{find_known_sequences}
\title{Given a genomic sequence it returns the relative coordinates where the sequence is known}
\usage{
find_known_sequences(x)
}
\arguments{
\item{x, }{a genomic sequence}
}
\value{
tibble
}
\description{
Given a genomic sequence it returns the relative coordinates where the sequence is known
}
