% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{GenomeBinning}
\alias{GenomeBinning}
\title{Bins the genome in equally sized bins}
\usage{
GenomeBinning(Chr_size, size = 5 * 10^6, Chr_filter = NULL, Cores = 1)
}
\arguments{
\item{Chr_size, }{a dataframe (chr=chrom name, size=chrom size)}

\item{size, }{bin size}

\item{Cores, }{number of threads to parallelise}
}
\value{
GenomicRanges object
}
\description{
Bins the genome in equally sized bins
}
