% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HeLa_Reference}
\alias{HeLa_Reference}
\title{HeLa S3 Referene RT}
\format{
tibble with 2831129 rows and 4 columns.
\describe{
\item{chr}{Chromosome position of a bin}
\item{start}{Start position of a bin}
\item{end}{End position of a bin}
\item{RT}{Replication timing data for a certain bin}
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSM923449}
}
\usage{
HeLa_Reference
}
\description{
A \code{tibble} containing bulk replication timing data derived from a Hg19 -> Hg38 liftover.
}
\keyword{datasets}
