% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HeLa_rep2}
\alias{HeLa_rep2}
\title{SingneCell RT Test object}
\format{
A list containing two tibbles variables:
\describe{
\itemize{
\item{PerCell}{Per Cell metrics: a tibble containing 293 rows and 6 columns}
\itemize{
\item{Cell}{Single Cell identifier}
\item{normalized_dimapd}{normalized Depth Independent Median Absolute deviation of Pairwise Differences}
\item{mean_ploidy}{Single Cell mean ploidy}
\item{ploidy_confidence}{Confidence in the CN calling}
\item{is_high_dimapd}{High DIMAPD cells have high internal variability and could be S-phase cells}
\item{is_noisy}{noisy cells have either poor ploidy confidence or high DIMAPD}
\item{coverage_per_1Mbp}{Reads coverage per megabase}
\item{basename}{Sample identifier}
\item{group}{Experimental group identifier}
}
}
\itemize{
\item{CNV}{Single cell copy number 310431 rows and 8 columns}
\itemize{
\item{Cell}{Single Cell identifier}
\item{chr}{Chromosome position of a bin}
\item{start}{Start position of a bin}
\item{end}{End position of a bin}
\item{copy_number}{Copy number Called for a bin}
\item{reads}{Normalised read count inside a bin}
\item{basename}{Sample identifier}
\item{group}{Experimental group identifier}
}
}
}
}
\source{
\url{https://www.nature.com/articles/s41467-022-30043-x}
}
\usage{
HeLa_rep2
}
\description{
A \code{list} with two \code{tibble}: PerCell and CNV for HeLa S3 cell replicate 2. Assembly Hg38.
}
\keyword{datasets}
