% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{IdAdapters}
\alias{IdAdapters}
\title{Identifies adapters present in a fastq}
\usage{
IdAdapters(File, Adapter_sequences)
}
\arguments{
\item{File, }{path to a fastq file}

\item{Adapter_sequences, }{a vector containing the adapters to test}
}
\description{
Identifies adapters present in a fastq
}
