% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{mergefastq}
\alias{mergefastq}
\title{merges a list of fastq files into one}
\usage{
mergefastq(
  FileList,
  NewName = NULL,
  output = file.path(getwd(), "mergedFastq")
)
}
\arguments{
\item{FileList, }{a list of files to merge}

\item{NewName, }{output file name if not provided this function looks for a common base between files}

\item{output, }{output directory}
}
\description{
merges a list of fastq files into one
}
