% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{mutate_sequence}
\alias{mutate_sequence}
\title{Randomly mutates the 4 nucleotides}
\usage{
mutate_sequence(Base)
}
\arguments{
\item{Base, }{either a list of characters or integers corresponding to the 4 nucleotides (A,T,C,G)}
}
\value{
vector
}
\description{
Randomly mutates the 4 nucleotides
}
