% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{Plot_bulkRT}
\alias{Plot_bulkRT}
\title{bulk RT plots of a region of interest}
\usage{
Plot_bulkRT(
  ...,
  Coordinates = list(chr = "chr1", start = 0, end = Inf),
  plotting_groups = NULL,
  sample_colors = NULL,
  highlight_regions = NULL
)
}
\arguments{
\item{..., }{one or multiple dataframes containing bulk and pseudo-bulk RTs and the following columns:chr, start, end, group, basename, RT}

\item{Coordinates, }{named list containing chr,start and end position}

\item{plotting_groups, }{optional: if provided samples can be plot together for comparison. It is a named list of arrays where the name is a sample basename if unique, otherwise basename - group, and the array contains the group(s) in which the sample has to be plot.}

\item{sample_colors, }{a vector, named or not, for each sample. If multiple samples have the same name use "basename - group" otherwise only basename}

\item{highlight_regions, }{optional: a list of arrays. each array contains start and end position of the region to highlight and optionally the color to use in third position}
}
\value{
plot
}
\description{
bulk RT plots of a region of interest
}
