% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{Post_processing}
\alias{Post_processing}
\title{Launch Kronos scRT post-processing}
\usage{
Post_processing()
}
\value{
Launches the shiny application
}
\description{
Main function to launch Kronos scRT post-processing in your browser.
}
