% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BinRepProbPlot.R
\name{Prepare_G1G2_phase_cells_forBinRepProb}
\alias{Prepare_G1G2_phase_cells_forBinRepProb}
\title{Prepare G1/G2 phase cells for BinRepProb plots}
\usage{
Prepare_G1G2_phase_cells_forBinRepProb(
  G1.G2,
  RT,
  quantile.range = c(0.25, 0.75)
)
}
\arguments{
\item{G1.G2, }{G1/G2 single cell CNV dataframe}

\item{RT, }{pseudo bulk replication timing dataframe}

\item{quantile.range, }{Ploidy quantile range to keep a G1/G2 cell}
}
\description{
Prepare G1/G2 phase cells for BinRepProb plots
}
