% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{Processing}
\alias{Processing}
\title{Launch Kronos scRT processing}
\usage{
Processing()
}
\value{
Launches the shiny application
}
\description{
Main function to launch Kronos scRT processing in your browser.
}
