% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{pseudoBulkRT}
\alias{pseudoBulkRT}
\title{calculates pseudobulk RT}
\usage{
pseudoBulkRT(S_scCN)
}
\arguments{
\item{S_scCN, }{scCN dataframe from Rebin function (after filtering,optional)}
}
\value{
tibble
}
\description{
calculates pseudobulk RT
}
