% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{Rebin}
\alias{Rebin}
\title{rebins kronos CN data (preprocessing output)}
\usage{
Rebin(PerCell, scCN, Bins, Sphase = NULL)
}
\arguments{
\item{PerCell, }{PerCell dataframe created by CallCNV}

\item{scCN, }{scCN dataframe created by CallCNV}

\item{Bins, }{a GenomicRanges object with genomic binning}

\item{Sphase, }{either True or False}
}
\value{
tibble
}
\description{
rebins kronos CN data (preprocessing output)
}
