% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{RebinRT}
\alias{RebinRT}
\title{rebins RT data}
\usage{
RebinRT(RT, Bins, Chr_filter = NULL, Basename = NULL, Group = NULL)
}
\arguments{
\item{RT, }{RT dataframe (chr,start,end,RT)}

\item{Bins, }{a GenomicRanges object with genomic binning}

\item{Chr_filter, }{filters chromosomes to keep in the analisys}
}
\value{
tibble
}
\description{
rebins RT data
}
