% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{recover_and_mutate}
\alias{recover_and_mutate}
\title{Given a genomic start and end recover a genomic sequence and simulate mutations due to sequencing}
\usage{
recover_and_mutate(Reads, Sequence, errorRate)
}
\arguments{
\item{Reads, }{tibble with start and end coordinates}

\item{Sequence, }{genomic sequence from which recover reads}

\item{errorRate, }{error rate to simulate mutation}
}
\value{
tibble
}
\description{
Given a genomic start and end recover a genomic sequence and simulate mutations due to sequencing
}
