% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{rejoinSubpop}
\alias{rejoinSubpop}
\title{reverts extractSubpop}
\usage{
rejoinSubpop(scCN, scRT, scVariability, subpopulation, RefRT = NULL)
}
\arguments{
\item{scCN, }{S-phase cells scCN dataframe from extractSubpop}

\item{scRT, }{pseudo-bulk-RT dataframe from extractSubpop}

\item{scVariability, }{single cell variability dataframe created by extractSubpop}

\item{subpopulation, }{dataframe containing sub-population info (Cell,basename,group,subpopulation)}

\item{RefRT, }{Reference RT of a group created by extractSubpop}
}
\value{
list
}
\description{
reverts extractSubpop
}
