% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{Replication_state}
\alias{Replication_state}
\title{binarizes data into replicated or not replicated bins}
\usage{
Replication_state(Samples, background, Chr_filter = NULL, cores = 3)
}
\arguments{
\item{Samples, }{a dataframe containing scCN info with the following columns:chr, start, end, group, basename, CN, index}

\item{background, }{a dataframe containing the median scCN info of the G1/G2 population with the following columns:chr, start, end, group, basename, background}

\item{Chr_filter, }{filters chromosomes to keep in the analysis}

\item{cores, }{number of threads to use while to parallelise}
}
\value{
tibble
}
\description{
binarizes data into replicated or not replicated bins
}
