% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{reshape_and_save}
\alias{reshape_and_save}
\title{Reshapes data into fastq format and save them into a file}
\usage{
reshape_and_save(Reads, Chrom, ID, path = "./")
}
\arguments{
\item{Reads, }{output of recover_and_mutate}

\item{ID, }{read identifier}

\item{path, }{path where to save the file}

\item{Chr, }{Chromosome from which the reads are originated}
}
\description{
Reshapes data into fastq format and save them into a file
}
