% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{rev_com}
\alias{rev_com}
\title{Given a genomic sequence it calculates its reverse complementary}
\usage{
rev_com(sequence)
}
\arguments{
\item{sequence, }{a genomic sequence}
}
\value{
string
}
\description{
Given a genomic sequence it calculates its reverse complementary
}
