% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RT_changes_plot.R
\name{RT_changes_plot}
\alias{RT_changes_plot}
\title{RT changes in function of ΔRT}
\usage{
RT_changes_plot(..., deltaRT = NULL, colors = NULL)
}
\arguments{
\item{..., }{one or more dataframes containing bulk and pseudo-bulk RTs and the following columns:chr, start, end, group, basename, RT}

\item{deltaRT, }{if deltaRT is set, the function behavior changes and it shows RT changes at that deltaRT threshold}

\item{colors, }{plot colors}
}
\value{
ggplot element
}
\description{
RT changes in function of ΔRT
}
