% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Clustering.R
\name{RT_clustering}
\alias{RT_clustering}
\title{Hclustering to identify RT signatures}
\usage{
RT_clustering(
  ...,
  deltaRT_th = 0.1,
  CrossingRT = T,
  n_clusters = NULL,
  colors = NULL
)
}
\arguments{
\item{..., }{bulk/pseudo-bulk RTs dataframes (chr,start,end,RT,group)}

\item{deltaRT_th, }{min RT change for a region to be considered}

\item{CrossingRT, }{IF true, only changes that are crossing the mid RT value will be considered}

\item{n_clusters, }{number of clusters}

\item{colors, }{min, midpoint and max colors}
}
\value{
list
}
\description{
Hclustering to identify RT signatures
}
