% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{scCNplot}
\alias{scCNplot}
\title{scCN plot}
\usage{
scCNplot(
  S_scCN,
  G_scCN,
  CN_limit = 4,
  Coordinates = NULL,
  rasterized_heatmap = F,
  heatmap_color_palette = c("viridis", "inferno", "magma", "plasma", "cividis"),
  manual_colors = NULL
)
}
\arguments{
\item{S_scCN, }{S-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{G_scCN, }{G1/G2-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{CN_limit, }{Max number of different level to visulize from the lowest one.}

\item{Coordinates, }{data frame containing chr,start and end position (multiple regions allowed, optional)}

\item{rasterized_heatmap, }{logic: whether or not the heatmap should be rasterized}

\item{heatmap_color_palette, }{palette choice (viridis, inferno, magma, plasma or cividis)}

\item{manual_colors, }{manual colors, if provided they overwrite heatmap_color_palette}
}
\value{
plot
}
\description{
scCN plot
}
