% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{scRTplot}
\alias{scRTplot}
\title{scRT plot of a region of interest}
\usage{
scRTplot(
  pseudoBulkRT,
  S_scCN,
  Coordinates,
  Plot = "scRT",
  rasterized_heatmap = F
)
}
\arguments{
\item{pseudoBulkRT, }{pseudoBulkRT that can be rbinded with a refrence RT (generatede bu RebinRT)}

\item{S_scCN, }{S-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{Coordinates, }{named list containing chr,start and end position}

\item{rasterized_heatmap, }{logic: whether or not the heatmap should be rasterized}
}
\value{
plot
}
\description{
scRT plot of a region of interest
}
