% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromFastqToBam.R
\name{sequential_trimming}
\alias{sequential_trimming}
\title{Sequentially trim reads}
\usage{
sequential_trimming(File, Read_adaters, outputdir, min_size = 25, cores = 1)
}
\arguments{
\item{File, }{path to a fastq file}

\item{outputdir, }{output directory}

\item{Adapter_sequences, }{a vector containing the adapters to test}
}
\description{
Sequentially trim reads
}
