% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{shell_interface}
\alias{shell_interface}
\title{Info to set Kronos scRT for command line}
\usage{
shell_interface(os = c("auto", "unix", "window"))
}
\arguments{
\item{os}{}
}
\description{
Info to set Kronos scRT for command line
}
