% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{split_into_categoreis}
\alias{split_into_categoreis}
\title{given an RT value (Early == 1 and Late == 0), this function returns its RT category base on the number of categories to return (2,3 or 5)}
\usage{
split_into_categoreis(RT, number = 2)
}
\arguments{
\item{RT, }{a number of a vector of numbers}

\item{number, }{number of categories to create}

\item{..., }{other parameters that can be passed to Rtsne}
}
\value{
a string or a vector of strings
}
\description{
given an RT value (Early == 1 and Late == 0), this function returns its RT category base on the number of categories to return (2,3 or 5)
}
