% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_subpop.R
\name{split_subpop}
\alias{split_subpop}
\title{Splits PerCell data frame based on identified subpopulations}
\usage{
split_subpop(PerCell, Supbop, group = NULL, basename = NULL)
}
\arguments{
\item{PerCell, }{Kornos PerCell data frame produced by Kronos}

\item{Supbop, }{Subpopulation grouping created by Kronos}
}
\value{
list
}
\description{
Splits PerCell data frame based on identified subpopulations
}
