% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/String_to_Range.R
\name{String_to_Range}
\alias{String_to_Range}
\title{converts a string representing a range of values into an array of integers}
\usage{
String_to_Range(x, div = ":")
}
\arguments{
\item{x, }{a string}

\item{div, }{symbol used to indicate a range}
}
\value{
an array of integers ranging from the first to the last number in the string
}
\description{
converts a string representing a range of values into an array of integers
}
