% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{T25_75}
\alias{T25_75}
\title{fit RT data into sigmoidal function and return T25 and T75}
\usage{
T25_75(df, sample, category)
}
\arguments{
\item{df, }{dataframe of group type}

\item{sample, }{the name of a sample}

\item{category, }{Genomic Category}
}
\value{
tibble
}
\description{
fit RT data into sigmoidal function and return T25 and T75
}
