% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10XtoKronos.R
\name{TenXtoKronos}
\alias{TenXtoKronos}
\title{It converts 10X Genomics data into Kronos format}
\usage{
TenXtoKronos(PerCell, CNV, basename, group)
}
\arguments{
\item{PerCell, }{PerCell 10XGenomics file path}

\item{CNV, }{CNV 10XGenomics file path}

\item{basename, }{sample basename}

\item{group, }{experimental group}
}
\value{
list
}
\description{
It converts 10X Genomics data into Kronos format
}
