% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TSNE.R
\name{TSNE}
\alias{TSNE}
\title{adds an option to set a seed in Rtsnee}
\usage{
TSNE(X, seed, ...)
}
\arguments{
\item{X, }{matrix; Data matrix (each row is an observation, each column is a variable)}

\item{seed, }{an integer}

\item{..., }{other parameters that can be passed to Rtsne}
}
\value{
Rtsne object
}
\description{
adds an option to set a seed in Rtsnee
}
