% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{TW_RTAnnotation}
\alias{TW_RTAnnotation}
\title{Annotate Variability dataframe based on RT Categories}
\usage{
TW_RTAnnotation(Variability, RT_Groups = 2)
}
\arguments{
\item{Variability, }{dataframe created by Kronos}

\item{RT_Groups, }{number of RT groups (1,2,3,5)}
}
\description{
Annotate Variability dataframe based on RT Categories
}
