% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{Twidth_barplot}
\alias{Twidth_barplot}
\title{Twidth barplotplot}
\usage{
Twidth_barplot(Variability, Twidth, Color = "lightblue", pval = NULL)
}
\arguments{
\item{Variability, }{dataframe created by either TW_RTAnnotation or TW_GenomeAnnotation}

\item{Twidth, }{dataframe created by Twidth}

\item{Color, }{barplot color(s)}

\item{pval, }{dataframe produced by twidth_pval}
}
\value{
ggplot element
}
\description{
Twidth barplotplot
}
