% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{Twidth_fit_data}
\alias{Twidth_fit_data}
\title{calculates twidth}
\usage{
Twidth_fit_data(df, ncores = 1)
}
\arguments{
\item{df, }{dataframe containing the columns group and category}

\item{cores, }{number of cores for parallelization}
}
\value{
tibble
}
\description{
calculates twidth
}
