% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{Twidth_pval}
\alias{Twidth_pval}
\title{Hypothesis test between different RT categories inside a sample}
\usage{
Twidth_pval(
  variability,
  twidth,
  pairs_to_test = NULL,
  adjust.methods = c("none", "holm", "hochberg", "hommel", "bonferroni", "BH", "BY",
    "fdr"),
  alternative = c("two.sided", "greater", "lower"),
  nIterations = 10^4,
  ncores = 3
)
}
\arguments{
\item{variability, }{dataframe created by Kronos}

\item{twidth, }{dataframe created by Twidth}

\item{pairs_to_test, }{a dataframe containing the columns Category1, Category2 to test}

\item{adjust.methods, }{correction method. Can be abbreviated. by default it is set as none}

\item{alternative, }{Hypothesis test. If pairs_to_test is null it is always two.sided}

\item{nIterations, }{number of iterations}

\item{ncores, }{number of cores for parallelization}
}
\description{
Hypothesis test between different RT categories inside a sample
}
