% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{Variability}
\alias{Variability}
\title{prepares variability file}
\usage{
Variability(S_scCN, scRT)
}
\arguments{
\item{S_scCN, }{scCN dataframe from Replication_state function (after filtering,optional)}

\item{scRT, }{pseudo-bulk-RT dataframe from pseudoBulkRT}
}
\value{
tibble
}
\description{
prepares variability file
}
