% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic.R
\name{WhoIsWho}
\alias{WhoIsWho}
\title{Assigns cell cycle staging metadata}
\usage{
WhoIsWho(PerCell, WhoIsWho)
}
\arguments{
\item{PerCell, }{per cell dataframe from CallCNV}

\item{WhoIsWho, }{dataframe containing two colums: Cell (character) and S_Phase (logical)}
}
\value{
tibble
}
\description{
Assigns cell cycle staging metadata
}
