% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{cat_levels}
\alias{cat_levels}
\title{Order in which RT categories should be plot}
\usage{
cat_levels(number)
}
\arguments{
\item{number, }{number of categories (2,3 or 5)}
}
\value{
character vector
}
\description{
Order in which RT categories should be plot
}
