% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Diagnostic.R
\name{diagnostic}
\alias{diagnostic}
\title{Helps identifying parameters for scRT analysis. It opends a shiny app}
\usage{
diagnostic(
  PerCell,
  interactive = T,
  min_RPMPH = 160,
  G1G2_th = NULL,
  S_th = NULL,
  First_half_factor = NULL,
  Second_half_factor = NULL,
  Automatic_correction = F,
  lunch.in.Rstudio = F,
  cores = 1
)
}
\arguments{
\item{PerCell, }{per cell dataframe from CallCNV}

\item{cores, }{number of cores to parallelise}
}
\value{
list
}
\description{
Helps identifying parameters for scRT analysis. It opends a shiny app
}
