% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{extractSubpop}
\alias{extractSubpop}
\title{separates S phase cells into subgroups and calculates relative pseudobulkRT}
\usage{
extractSubpop(scCN, scRT, scVariability, subpopulation, RefRT = NULL)
}
\arguments{
\item{scCN, }{Sphase cells scCN dataframe from Rebin function (after filtering,optional)}

\item{scRT, }{pseudobulkRT dataframe}

\item{scVariability, }{single cell variability dataframe the Variability function}

\item{subpopulation, }{dataframe containing subpopulation info (Cell,basename,group,subpopulation)}

\item{RefRT, }{Reference RT dataframe from RebinRT}
}
\value{
list
}
\description{
separates S phase cells into subgroups and calculates relative pseudobulkRT
}
