% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scRT_processing.R
\name{FilterCells}
\alias{FilterCells}
\title{filter cells}
\usage{
FilterCells(scCN, min_cor = 0.25, ApplyFilter = F)
}
\arguments{
\item{scCN, }{scCN dataframe from Replication_state function}

\item{min_cor, }{min correlation to be kept}
}
\value{
list
}
\description{
filter cells
}
