% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KCorr_plot.R
\name{KCorr_plot}
\alias{KCorr_plot}
\title{Customized GGally::ggparis function}
\usage{
KCorr_plot(df, method = "spearman")
}
\arguments{
\item{df, }{a dataframe containing bulk and pseudo-bulk RTs and the following columns:chr, start, end, group, basename, RT}
}
\value{
ggplot element
}
\description{
Customized GGally::ggparis function
}
