% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{scCNplot}
\alias{scCNplot}
\title{scCN plot}
\usage{
scCNplot(
  S_scCN,
  G_scCN,
  CN_limit = 4,
  Coordinates = NULL,
  rasterized_heatmap = F
)
}
\arguments{
\item{S_scCN, }{S-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{G_scCN, }{G1/G2-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{CN_limit, }{Max number of different level to visulize from the lowest one.}

\item{Coordinates, }{data frame containing chr,start and end position (multiple regions allowed, facultative)}

\item{rasterized_heatmap, }{logic: whether or not the heatmap should be rasterized}
}
\value{
plot
}
\description{
scCN plot
}
