% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plot_functions.R
\name{scRTplot}
\alias{scRTplot}
\title{scRT plot of a region of interest}
\usage{
scRTplot(
  pseudoBulkRT,
  S_scCN,
  Coordinates,
  Plot = "scRT",
  rasterized_heatmap = F,
  heatmap_colors = NULL,
  sample_colors = NULL
)
}
\arguments{
\item{pseudoBulkRT, }{pseudo-Bulk RT that can be rebinned with a reference RT (generated by RebinRT)}

\item{S_scCN, }{S-phase scCN dataframe created by Replication_state (optionally filtered)}

\item{Coordinates, }{named list containing chr,start and end position}

\item{rasterized_heatmap, }{logic: whether or not the heatmap should be rasterized}

\item{heatmap_colors, }{a vector of colors for the heatmap.(Replicated/min value and Unreplicated/max value)}

\item{sample_colors, }{a vector, named or not, for the samples}
}
\value{
plot
}
\description{
scRT plot of a region of interest
}
