% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Variability_functions.R
\name{Twidth_extended_plot}
\alias{Twidth_extended_plot}
\title{Extended Twidth plot}
\usage{
Twidth_extended_plot(Variability, Fitted_data, Twidth)
}
\arguments{
\item{Variability, }{dataframe created by either TW_RTAnnotation or TW_GenomeAnnotation}

\item{Fitted_data, }{dataframe creted by Twidth_fit_data}

\item{Twidth, }{dataframe created by Twidth}
}
\value{
ggplot element
}
\description{
Extended Twidth plot
}
